+++ CHANGELOG.md
@@ -63,6 +63,11 @@
 
 * Minor tweak to avoid potential dictionary trouble with double-adding - thanks [samartzidis]
 
+## 4.5.0
+
+* Add callback to allow for customizing the `IConnectionFactory` instance used by the transport
+* Make `RabbitMqOptionsBuilder` non-CLS compliant (necessary to be able to customize RabbitMQ's connection factory)
+
 
 ---